% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_condition_summary.R
\name{compute_condition_summary}
\alias{compute_condition_summary}
\title{Computes summary data for each experimental conition/cell.}
\usage{
compute_condition_summary(param_summary_data, facet_var, ind_vars)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
