% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_errorbar_lengths.R
\name{compute_errorbar_lengths}
\alias{compute_errorbar_lengths}
\title{Computes summary data for each experimental conition/cell.}
\usage{
compute_errorbar_lengths(
  exp_analytical_days,
  iv_level,
  num_measurements,
  sample_size
)
}
\arguments{
\item{exp_analytical_days}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
