% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_param_density_plot.R
\name{compute_param_hist_elements}
\alias{compute_param_hist_elements}
\title{density_lower_x' = density_lower_x, 'density_upper_x' = density_upper_x,}
\usage{
compute_param_hist_elements(base_density_plot, param_ind_data_ls)
}
\description{
density_lower_x' = density_lower_x, 'density_upper_x' = density_upper_x,
}
