% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_parameter_summary.R
\name{compute_parameter_summary}
\alias{compute_parameter_summary}
\title{Computes summary statistics for simulated experiment data.}
\usage{
compute_parameter_summary(data, exp_num)
}
\arguments{
\item{data}{experiment data}

\item{exp_num}{experimental conditions}
}
\description{
Computes summary statistics for simulated experiment data.
}
