% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_var_to_sd.R
\name{convert_var_to_sd}
\alias{convert_var_to_sd}
\title{Converts variances to standard deviations for random-effect and residual parameters.}
\usage{
convert_var_to_sd(param_summary_data)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Converts variances to standard deviations for random-effect and residual parameters.
}
