% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_factor_levels.R
\name{extract_factor_levels}
\alias{extract_factor_levels}
\title{Computes summary data for each experimental conition/cell.}
\usage{
extract_factor_levels(param_summary_data)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
