% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_summary_table.R
\name{generate_column_latex}
\alias{generate_column_latex}
\title{Computes summary data for each experimental condition/cell.}
\usage{
generate_column_latex(col_number, parameter_data, abs_cutoff = 0.4)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental condition/cell.
}
