% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_day_likert_facet_plot.R
\name{generate_day_likert_facet_plot}
\alias{generate_day_likert_facet_plot}
\title{Generate facetted plots}
\usage{
generate_day_likert_facet_plot(
  analytical_data,
  target_col = "measurement_spacing",
  target_value,
  x_axis_var,
  x_axis_name,
  exp_num,
  beta_lower,
  beta_upper,
  beta_ticks
)
}
\arguments{
\item{data}{experiment data}
}
\description{
Generate facetted plots
}
