% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_day_likert_data_sets.R
\name{generate_likert_days_data_sets}
\alias{generate_likert_days_data_sets}
\title{Computes summary data for each experimental conition/cell.}
\usage{
generate_likert_days_data_sets(summary_data, spacing, exp_num)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
