% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_param_density_plot.R
\name{generate_param_density_plot}
\alias{generate_param_density_plot}
\title{Computes summary data for each experimental conition/cell.}
\usage{
generate_param_density_plot(
  raw_exp_data,
  param_summary_data,
  spacing,
  num_measurements,
  sample_size
)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
