% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_facet_plot.R
\name{generate_summary_facet_plot}
\alias{generate_summary_facet_plot}
\title{Computes summary data for each experimental conition/cell.}
\usage{
generate_summary_facet_plot(
  condition_data,
  lower_y_limit,
  upper_y_limit,
  ticks,
  x_axis_var = "midpoint",
  x_axis_name = "Midpoint location (days)",
  facet_var = "measurement_spacing",
  exp_num,
  num_rows = 2,
  num_cols = 2,
  legend_position = c(0.5, 0.5),
  legend_direction = "horizontal",
  file_name,
  dodge_width = 0.8,
  point_size = 15,
  line_size = 2,
  error_bar_width = 0.8,
  error_bar_size = 2.5,
  y_axis_var = "estimate",
  y_axis_name = "Parameter Bias (Percentage Error)",
  grouping_var = "number_measurements",
  shape_legend_title = "Number of \\nMeasurements",
  panel_spacing_y = 12,
  dodge_position = position_dodge(width = 0.8)
)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
