% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_function.R
\name{generic_function}
\alias{generic_function}
\title{Computes summary data for each experimental conition/cell.}
\usage{
generic_function()
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
