% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bias_var_omega_table.R
\name{print_bias_var_omega_table}
\alias{print_bias_var_omega_table}
\title{Computes summary data for each experimental conition/cell.}
\usage{
print_bias_var_omega_table(
  exp_data,
  target_col,
  target_value,
  ind_vars = c("number_measurements", "midpoint"),
  ind_var_acronyms,
  caption = "$\\\\upomega^2$ Values for Manipulated Variables With Equal Spacing",
 
    footnote = "NM = number of measurements, M = population value set for $\\\\\\\\upbeta_{fixed}$,\\n                                       NM x M = interaction between number of measurements and population value set for $\\\\\\\\upbeta_{fixed}$",
  parameter_labels = c("$\\\\upbeta_{fixed}$ (Figure \\\\ref{fig:exp1_plot_equal}A)",
    "$\\\\upbeta_{random}$ (Figure \\\\ref{fig:exp1_plot_equal}B)",
    "$\\\\upgamma_{random}$ (Figure \\\\ref{fig:exp1_plot_equal}C)",
    "$\\\\upgamma_{random}$ (Figure \\\\ref{fig:exp1_plot_equal}D)")
)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\value{
Returns a data table.
}
\description{
Computes summary data for each experimental conition/cell.
}
