% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_convergence_table.R
\name{print_conv_table}
\alias{print_conv_table}
\title{Computes summary data for each experimental conition/cell.}
\usage{
print_conv_table(
  table_ready_data,
  caption_name,
  col_header_name,
  IV_names,
  column_names
)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
