% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_param_table.R
\name{print_param_table}
\alias{print_param_table}
\title{Computes summary data for each experimental conition/cell.}
\usage{
print_param_table(
  table_ready_data,
  parameter_name,
  caption_name,
  col_header_name,
  IV_names,
  column_names,
  footnote,
  first_col_width = "3cm"
)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
