% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_factors.R
\name{relevel_factors}
\alias{relevel_factors}
\title{Computes summary data for each experimental conition/cell.}
\usage{
relevel_factors(param_summary_data, factor_levels_df)
}
\arguments{
\item{param_summary_data}{parameter summary data (created from compute_parameter_summary)}
}
\description{
Computes summary data for each experimental conition/cell.
}
