% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Removes outliers from each generated data set.}
\usage{
remove_outliers(data, exp_conditions = locate_ivs(data))
}
\arguments{
\item{data}{experiment data}

\item{exp_conditions}{experimental conditions}
}
\description{
Removes outliers from each generated data set.
}
